app/requirements.txt
langchain==0.3.15
langchain-community==0.3.15
langchain-groq===0.2.3
unstructured==0.16.15
selenium==4.28.1
chromadb==0.6.3
streamlit==1.40.1
pandas==2.2.3
python-dotenv==1.0.1